requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
aiofiles>=23.2.1
boto3>=1.34.0
structlog>=24.1.0
pydantic>=2.7.1
pydantic-settings>=2.2.1
packaging>=24.0
jinja2>=3.1.3
pandas>=2.2
openpyxl>=3.1
orjson>=3.10
python-slugify>=8.0
tabulate>=0.9
httpx[http2]>=0.27
backoff>=2.2.1
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20
alembic>=1.13
passlib[bcrypt]>=1.7
pytest-asyncio>=0.23